ictfleet/backend/system_vehicule/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
python-decouple
PyJWT
requests
groq
openpyxl
